+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* ✅ test(slice-chunk): made test cases and benchmarks more accurate
 * 🎉 feat(slice-chunk): chunk function added, updated, and optimized
 
 .. 1.0.0 (yyyy-mm-dd)